Use mostly native int for variable types, because a small test shows that adding two ints takes less code than adding two shorts.
sizeof() gives these values (note that 'unsigned' is short for 'unsigned int'):

arch   bool char short int long longlong unsigned
armv6  1    1    2     4   4    8        4
amd64  1    1    2     4   8    8        4
i386   1    1    2     4   4    8        4

The binary size is now reduced by 1% to 2% and less casts are required.

Mostly (un)signedness was preserved, but some types were changed to 'int'
to save some now unnecessary casts.

There are no measurable speedups in dcf77pi however.

It also has the benefit of not having to include <stdint.h> which
include-what-you-use 0.7 does not fully understand on FreeBSD.

While here, push down some loop counter declarations.

+++ bits1to14.h
@@ -26,11 +26,10 @@ SUCH DAMAGE.
 #ifndef DCF77PI_BITS1TO14_H
 #define DCF77PI_BITS1TO14_H
 
-#include <stdint.h>
 struct GB_result;
 
 /** Length of the third-party buffer in bits */
-extern const uint8_t tpBufLen;
+extern const unsigned tpBufLen;
 
 /** Indicates the type of the third party contents. */
 enum eTP {
@@ -49,7 +48,7 @@ enum eTP {
  * @param bitpos The current bit position.
  * @param bit The current bit information.
  */
-void fill_thirdparty_buffer(uint8_t minute, uint8_t bitpos,
+void fill_thirdparty_buffer(unsigned minute, unsigned bitpos,
     const struct GB_result * const bit);
 
 /**
@@ -57,7 +56,7 @@ void fill_thirdparty_buffer(uint8_t minute, uint8_t bitpos,
  *
  * @return The third party buffer.
  */
-const uint8_t * const get_thirdparty_buffer(void);
+const unsigned * const get_thirdparty_buffer(void);
 
 /**
  * Retrieve the type of the third party contents.
