+++ config.c
@@ -26,7 +26,6 @@ SUCH DAMAGE.
 #include "config.h"
 
 #include <errno.h>
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -38,9 +37,9 @@ static const char *key[] = {
 
 #define NUM_KEYS (sizeof(key) / sizeof(key[0]))
 
-static const uint8_t max_key_len = 20;
-static const uint8_t max_val_len = 255;
-static const uint16_t max_len = (max_key_len + 3 + max_val_len + 2);
+static const unsigned max_key_len = 20;
+static const unsigned max_val_len = 255;
+static const unsigned max_len = (max_key_len + 3 + max_val_len + 2);
     /* "k = v\n\0" */
 
 static char *value[NUM_KEYS];
