+++ decode_time.h
@@ -27,7 +27,6 @@ SUCH DAMAGE.
 #define DCF77PI_DECODE_TIME_H
 
 #include <stdbool.h>
-#include <stdint.h>
 struct tm;
 
 enum eDT_length {
@@ -146,7 +145,7 @@ void init_time(void);
  * @return A structure containing the results of all the checks performed on
  *   the calculated time.
  */
-const struct DT_result * const decode_time(uint8_t init_min, uint8_t minlen,
-    uint32_t acc_minlen, const uint8_t * const buffer, struct tm * const time);
+const struct DT_result * const decode_time(unsigned init_min, unsigned minlen,
+    unsigned acc_minlen, const int * const buffer, struct tm * const time);
 
 #endif
